Backend/src/test/java/com/cd/controller/UserControllerTest.java
package com.cd.controller;

import com.cd.dto.UserInputDto;
import com.cd.dto.UserOutputDto;
import com.cd.service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.*;

class UserControllerTest {

    @InjectMocks
    private UserController userController;

    @Mock
    private UserService userService;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void createUser_ShouldReturnUserInputDto() {
        UserInputDto userInputDto = new UserInputDto();
        when(userService.createUser(any(UserInputDto.class))).thenReturn(userInputDto);

        UserInputDto result = userController.createUser(userInputDto);

        assertNotNull(result);
        verify(userService, times(1)).createUser(userInputDto);
    }

    @Test
    void getAllUsers_ShouldReturnListOfUsers() {
        when(userService.getAllUsers()).thenReturn(Collections.emptyList());

        List<UserOutputDto> result = userController.getAllUsers();

        assertNotNull(result);
        verify(userService, times(1)).getAllUsers();
    }

    @Test
    void deleteUser_ShouldReturnNoContent() {
        doNothing().when(userService).deleteUser(1);

        ResponseEntity<Void> response = userController.deleteUser(1);

        assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
        verify(userService, times(1)).deleteUser(1);
    }
}
